+++ changelog.md
@@ -1,4 +1,5 @@
-# Ekang API - Changelog
+# Si-Ekang API - Changelog
+
 All notable changes to this project will be documented in this file.
 
 > The format is based on
@@ -7,6 +8,7 @@ All notable changes to this project will be documented in this file.
 > [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Types of changes
+
 - `Added` for new features.
 - `Changed` for changes in existing functionality.
 - `Deprecated` for soon-to-be removed features.
@@ -16,23 +18,46 @@ All notable changes to this project will be documented in this file.
 
 ## Work in progress
 
-## Version 1.1.0 - 2022/01/17
+### Added
+
+- Route `GET /categories` for retrieving all categories.
+- Route `GET /categories/{id}/translations` for retrieving all category's
+  translations.
+
 ### Changed
+
+- Update paginated translations retrieving by adding an `id` to each of them.
+
+### Removed
+
+- OpenAPI documentation.
+
+## Version 1.1.0 - 2022/01/17
+
+### Changed
+
 - Upgrade Gradle to 7.3.1.
 
 ## Version 1.0.1 - 2021/12/02
+
 ### Fixed
+
 - Upgrade base routing version to `v1`.
 
 ## Version 1.0.0 - 2021/12/02
+
 ### Added
+
 - Create the route `GET /translations` for retrieving paginated translations.
 - Tag for grouping operations related to translations in the documentation.
 
 ## Version 0.1.0 - 2021/11/08
+
 ### Added
+
 - Create a new server with [Ktor](https://ktor.io/).
 - Create [OpenAPI](https://swagger.io/specification/) documentation.
 
 ### Security
+
 - Allow HTTP requests from specific origins.